% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicRegionsFromGeneSymbols.r
\name{getGenomicRegionsFromGeneSymbols}
\alias{getGenomicRegionsFromGeneSymbols}
\title{Function that reads in a list of probesets' ID's for the GPL96 platform and returns the annotations for them, if found}
\usage{
getGenomicRegionsFromGeneSymbols(
  thisGeneSymbolList,
  thisTrackName,
  saveOutputFileFlag,
  outputFolderName,
  outputFileFormat,
  verboseFlag
)
}
\arguments{
\item{thisGeneSymbolList}{list of gene symbols}

\item{thisTrackName}{name of the track to print in the file}

\item{saveOutputFileFlag}{says if the output file should be saved or not}

\item{outputFolderName}{name of the output file directory}

\item{outputFileFormat}{parameter that specifies the file format of the output}

\item{verboseFlag}{if TRUE, intermediate messages are going to be printed}
}
\value{
a dataframe with the annotations containing the chromosome regions
}
\description{
Function that reads in a list of probesets' ID's for the GPL96 platform and returns the annotations for them, if found
}
\examples{
gene_symbols_list <- c("AK4", "ALDOC", "EGLN1", "FAM162A", "MTFP1", "PDK1", "PGK1")
verboseFlag <- FALSE
trackName <- "neuroblastomaPrognosticSignature2020"
outputFolder  <- "./test"
outputFileFormat <- "BED"
outputFileFlag <- FALSE
arrangedAnnotations <- getGenomicRegionsFromGeneSymbols(gene_symbols_list, trackName, outputFileFlag, outputFolder, outputFileFormat, verboseFlag)
}
