% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicRegionsFromGeneSymbols.r
\name{jetsetCheckLookup}
\alias{jetsetCheckLookup}
\title{Function that peforms lookup check from jetset}
\usage{
jetsetCheckLookup(query, result)
}
\arguments{
\item{query}{query}

\item{result}{result}
}
\value{
nothing
}
\description{
Function that peforms lookup check from jetset
}
\examples{
gene_symbols_list1 <- c("AK4", "ALDOC", "EGLN1", "FAM162A", "MTFP1", "PDK1", "PGK1")
gene_symbols_list2 <- c("AK4", "ALDOC", "EGLN1", "FAM162A", "MTFP1", "PDK1", "PGK1")
jetsetCheckLookup(gene_symbols_list1, gene_symbols_list2)
}
