% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicRegionsFromGeneSymbols.r
\name{jmapHgu133aReimplemented}
\alias{jmapHgu133aReimplemented}
\title{Function that maps gene symbols to probesets of the hgu133a platform from jmap}
\usage{
jmapHgu133aReimplemented(symbols_list)
}
\arguments{
\item{symbols_list}{list of gene symbols}
}
\value{
a list of probesets related to the input gene symbols
}
\description{
Function that maps gene symbols to probesets of the hgu133a platform from jmap
}
\examples{
gene_symbols_list <- c("AK4", "ALDOC", "EGLN1", "FAM162A", "MTFP1", "PDK1", "PGK1")
mappings <- (gene_symbols_list)
}
